feat: add 'none' filter
To allow force application of KMS changes recursively.
